requirements.txt
opencv-python
paddlenlp